Add admin panel to add/update sets. Fixed other issues

+++ NerdBot/NerdBot_DatabaseUpdater/MtgData/MtgJsonCard.cs
@@ -11,7 +11,7 @@ namespace NerdBot_DatabaseUpdater.MtgData
         public string Layout { get; set; } // Contains layout of card, e.g. layout: "token" or layout: "normal"
         public string Name { get; set; }
         public string MultiverseId { get; set; }
-        public int CMC { get; set; }
+        public double CMC { get; set; }
         public string[] Colors { get; set; }
         public string[] Types { get; set; }
         public string[] SubTypes { get; set; }
